move source files to parent directory
