+++ spec/gl-matrix/mat4-spec.js
@@ -1,6 +1,6 @@
-var mat4 = require("../../src/gl-matrix/mat4.js");
-var vec3 = require("../../src/gl-matrix/vec3.js");
-var quat = require("../../src/gl-matrix/quat.js");
+var mat4 = require("../../src/mat4.js");
+var vec3 = require("../../src/vec3.js");
+var quat = require("../../src/quat.js");
 
 function buildMat4Tests() {
     return function() {
