+++ spec/gl-matrix/quat-spec.js
@@ -1,8 +1,8 @@
 describe("quat", function() {
-    var mat3 = require("../../src/gl-matrix/mat3.js");
-    var mat4 = require("../../src/gl-matrix/mat4.js");
-    var quat = require("../../src/gl-matrix/quat.js");
-    var vec3 = require("../../src/gl-matrix/vec3.js");
+    var mat3 = require("../../src/mat3.js");
+    var mat4 = require("../../src/mat4.js");
+    var quat = require("../../src/quat.js");
+    var vec3 = require("../../src/vec3.js");
 
     var out, quatA, quatB, result;
     var vec, id, deg90;
