+++ src/gl-matrix.js
@@ -1,9 +1,9 @@
-exports.glMatrix = require("./gl-matrix/common.js");
-exports.mat2 = require("./gl-matrix/mat2.js");
-exports.mat2d = require("./gl-matrix/mat2d.js");
-exports.mat3 = require("./gl-matrix/mat3.js");
-exports.mat4 = require("./gl-matrix/mat4.js");
-exports.quat = require("./gl-matrix/quat.js");
-exports.vec2 = require("./gl-matrix/vec2.js");
-exports.vec3 = require("./gl-matrix/vec3.js");
-exports.vec4 = require("./gl-matrix/vec4.js");
+exports.glMatrix = require(".common.js");
+exports.mat2 = require(".mat2.js");
+exports.mat2d = require(".mat2d.js");
+exports.mat3 = require(".mat3.js");
+exports.mat4 = require(".mat4.js");
+exports.quat = require(".quat.js");
+exports.vec2 = require(".vec2.js");
+exports.vec3 = require(".vec3.js");
+exports.vec4 = require(".vec4.js");
