+++ src/quat.js
